fix: imperative restart should rollout the primary pod (#7122)
This patch updates the operator to create a new Pod for the primary
instance when an imperative restart is requested.

Previously, if `primaryUpdateMethod` was set to `restart`, the operator
would recreate the replica Pods but only perform an in-place restart of
the primary Pod.

This approach could lead to inconsistencies between the definitions of
the primary and the replica Pods, as only the replicas were fully
recreated.

Closes #7120

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>