feat(cmd,plugin): make `hibernate` command declarative (#7155)
## Release Notes

The `hibernate on/off` commands in the `cnpg` plugin for `kubectl` now
serve as shortcuts for the declarative hibernation procedure. The
previously available imperative implementation has been removed in
favour of the declarative approach. Additionally, the `hibernate status`
command has been deprecated in favour of the standard `status` command,
which now provides a clearer interface for hibernated clusters.

**Users should not upgrade to version 1.26 of both the plugin and the
operator unless they are prepared to migrate to the declarative
method.**

Closes #6870

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>