feat: allow to bypass the validation webhook
This patch enables users to bypass the validation webhook by adding
an annotation called `cloudnative-pg.io/validation` with the
value `enabled`.

This feature is necessary for making unrestricted changes to the
cluster definition, which allows users to perform tasks that the
operator typically restricts, such as reducing the size of volumes
and rebuilding them individually.

However, this added flexibility also increases the risk of executing
potentially harmful operations.

Fixes: #7117

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>