feat: limit backup permissions to involved objects
This patch modifies the operator to explicitly enumerate the backup
objects that the instance manager needs to interact with, rather than
granting it permission to access all backup objects within the
containing namespace.

This change improves our stance on the principle of least privilege,
limiting access to only the necessary resources.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>