test: cleanup leftover logs from db management E2Es (#7217)
Closes #6411 

Usually, at the end of a green test, `cleanupNamespace()` will take care
of deleting the testing logs captured and issuing an `objects.Delete`
request (we do not actually wait for the namespace to be gone though).

This Database management E2E is an exception to the usual way we
create/destroy testing namespaces.
Here we want to test that Database object finalizers do not prevent the
deletion of a namespace, hence we need to ensure the namespace is
properly gone before declaring that the test succeded. After doing that,
we now cleanup the testing logs.

Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>