feat: allow bypassing the validation webhook (#7196)
This patch introduces the `cnpg.io/validation` annotation, enabling
users to disable the validation webhook on CloudNativePG-managed
resources.

This capability is essential for making unrestricted modifications to
the cluster definition, allowing operations that the operator would
typically prevent—such as reducing volume sizes and rebuilding them
individually.

However, this increased flexibility comes with the risk of executing
potentially harmful operations, so it should be used with caution.

Fixes: #7117

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>