feat: declarative offline in-place major upgrades of PostgreSQL (#6664)
This PR introduces the ability to upgrade PostgreSQL clusters to a
higher major version, enabling a fully declarative approach for major
version upgrades within CloudNativePG.

**IMPORTANT:** While the upgrade job area still requires refactoring,
we’ve decided to merge this patch into the main branch and release
version 1.26.0-rc1 to gather early community feedback at KubeCon Europe
2025 in London. The refactoring will likely require a second release
candidate before the final 1.26.0 release.

With this update, users can specify a new image with a higher major
version using any supported method. The operator tracks the previous
image version in `.status.majorVersionUpgradeFromImage`, which also
serves as an indicator that a major version upgrade is in progress.

The upgrade process starts by deleting all pods and jobs to ensure the
data directory is inaccessible. A dedicated upgrade job is created for
the primary instance, using an `initContainer` to copy the old
PostgreSQL installation and running `pg_upgrade --link` for an efficient
migration. Once the upgrade completes, all non-primary instances are
removed, and new standbys are cloned to restore the desired number of
instances.

Closes #4682 

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>