ci: add digest to operator manifest for testing (#7348)
Given that we produce multiple images with the same tags for testing,
the upgrade tests can fail now that ImagePullPolicy is set to Always.
Fix the digest in the manifest to properly specify the image that is
used.

Closes #7347

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
(cherry picked from commit c7432d793fb5b1e1b44280ee24fa12e37080ae89)