fix: replace MajorVersionUpgradeFromImage with PGDataImageInfo in status (#7403)
Include both the major version and the image used to run the data
directory in the status. This addresses a failure scenario during major
upgrades when using an ImageCatalog.

If the images in the catalog do not follow the CloudNativePG tag
conventions, the operator cannot reliably determine the starting major
version, causing the upgrade to fail. By explicitly storing both the
image and its associated major version, we ensure the upgrade process
has the necessary context.

Closes #7391

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>