test: allow configuring a custom image repository in major upgrade E2Es (#7303)
This patch goes together with
https://github.com/cloudnative-pg/postgres-trunk-containers/pull/80.
The goal is to allow configuring a custom repository (in this case the
`postgresql-trunk`) instead of the default ones (via an env variable),
which will be used to fetch the target images required to perform the
major upgrade scenarios.
This will allow us to test major upgrades to development versions of
PG18.

Closes #7302


Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>