chore: move dataDurability defaulting logic to webhook to avoid OLM issues (#7600)
Defining an OpenAPI default for a field within an optional section (such
as `spec.postgresql.synchronous`) causes OLM to treat the entire section
as required. This leads to validation errors when other mandatory fields
in that section are not set.

To resolve this, the defaulting logic for
`spec.postgresql.synchronous.dataDurability` has been moved from the CRD
schema to the defaulting webhook.

Closes #7599

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>