test(e2e): strenghten readiness checks (#7603)
Fixes some cases in the e2e where the tests would race to the next test
when the operator is not ready yet, or when the nodes are not ready.

Refactor the code to improve reuse and simplify signatures.

Closes #7589

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
(cherry picked from commit 67b8594ac18c37c46ea5a3b4c9d373fc2492f9d3)