feat: "unrecoverable" annotation
This patch make the operator look for the "alpha.cnpg.io/unrecoverable=true"
annotation in the PG instance Pods.
If such an annotation is detected, the operator will destroy the PVCs
and the Pod itself and recreate it from the other instances.

This annotation is available only for replica instances.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>