feat: user maps for fixed users (#7697)
Currently it's not possible to use self-managed client certificates in environments that don't allow issuing certs with just`streaming_replica` as the common name.
Because those entries in pg_hba take priority, it's just not possible to work around it - this user is hardcoded in the cluster replica join logic.

With a default user map it becomes possible to remap any cert to allow the streaming replica user access.

Closes #7697

Signed-off-by: Szymon Soloch <ssoloch@opera.com>