feat: keep track of the PG system ID in the status (#7717)
This patch add field names `.status.SystemID`, that keeps track of the
system ID reported by the PostgreSQL instances.

The new field is set only if the reported system ID is consistent across
all the instances. If there is no reported system ID or if there are
inconsistencies, the field set to empty.

A new condition keeps track of the consistency status.

Closes: #7716

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>