fix(probes): avoid liveness probe to timeout when not needed
When the API server is not reachable, the default liveness probe
configuration may trigger an automatic shutdown of the PG instance, even
when the isolation checker is not active.

This happens bacause a call to the API server may time out after the
time out of the Kubelet.

This fix includes a safe limit, choosen after the default timeout of
Pods' probes and is safe to use with the default CNPG configuration.

Closes: #7901

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>