fix(probes): avoid liveness probe to timeout when not needed (#7902)
When the API server is not reachable, the default liveness probe
configuration may trigger an automatic shutdown of the PG instance, even
when the isolation checker is not active.

This happens because a call to the API server may time out after the
time out of the Kubelet.

This fix includes a safe limit, chosen after the default timeout of
Pods' probes and is safe to use with the default CNPG configuration.

Closes: #7901

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
(cherry picked from commit 41cbd96bf09676f0f9ac545d55a8f12db5339dc6)