fix(cluster,backup): ensure Backup status is set after failure (#7898)
This patch ensures that the status stanza of the Backup resource is
set after a backup has failed, including the `method` field.

As a consequence of this, the `LastBackupSucceeded` condition and
the `lastFailedBackup` field will be properly set on the Cluster resource.

Closes #7892

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
(cherry picked from commit b21f768be78f18cb79790d69363a3af60a79e55b)