feat(spec,probes): primary isolation check in the liveness probe (#7845)
This patch promotes the liveness pinger feature from its previous
experimental status (using the `alpha.cnpg.io/livenessPinger`
annotation) to a stable, supported configuration within the `Cluster`
resource, located in the `.spec.probes.liveness.isolationCheck` section.

The feature is now called "isolation check" and is enabled by default in
the liveness probe. The values for `requestTimeout` and
`connectionTimeout` are set to 1 second by default.

Closes #7551 

### Release notes

```
Primary isolation checks for the liveness probe: the liveness pinger, previously configured with the `alpha.cnpg.io/livenessPinger` annotation, is now included in the `Cluster` specification under the `isolationCheck` stanza and is enabled by default, enforcing the shutdown of an isolated primary.
```

---------

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>