fix(backup,online): fail backup if targetPod becomes unhealthy (#7944)
This patch ensures that the backup process is immediately aborted if the
targetPod becomes unhealthy during an online backup operation.
Previously, the backup would continue even if the target instance was no
longer in a healthy state, potentially resulting in incomplete or
inconsistent backups. By monitoring the pod's health status, the process
now fails fast, providing early feedback and increasing the accuracy of
the Backup object status.

Related #7503
Closes #7905

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>