fix(postgres,instance): set postgres workers `oom_score_adj` (#7891)
When the quality of service is set to "Guaranteed", CloudNativePG sets
the `PG_OOM_ADJUST_VALUE` for the `postmaster` process to `0`.

This allows the `postmaster` to retain its low Out-Of-Memory (OOM) score
of `-997`, while its child processes run with an OOM score adjustment of
`0`. As a result, if the OOM killer is triggered, it will terminate the
child processes before the `postmaster`.

This behaviour helps keep the PostgreSQL instance alive for as long as
possible and enables a clean shutdown procedure in the event of an
eviction.

Closes #7132

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>