chore: azurite test failing due to API versions mismatch
The azurite tests were failing with the following error:
ERROR: Can't connect to cloud provider: The API version
2025-07-05 is not supported by Azurite. Please upgrade
Azurite to latest version and retry. If you are using
Azurite in Visual Studio, please check you have installed
latest Visual Studio patch. Azurite command line parameter
\"--skipApiVersionCheck\" or Visual Studio Code configuration
\"Skip Api Version Check\" can skip this error.

Added the `--skipApiVersionCheck` to the azurite server we
run and that avoid having this issue.

This was already reported in the azurite community here:
https://github.com/azure/azurite/issues/2564

Closes #8059

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>