docs(cnpg-i)!: deprecate backup fields in `Cluster` status
Deprecates `firstRecoverabilityPoint`, `firstRecoverabilityPointByMethod`,
`lastSuccessfulBackup`, `lastSuccessfulBackupByMethod`, and `lastFailedBackup`
in `Cluster.status` due to the transition to a plugin-based, backup-agnostic
architecture in CloudNativePG.

BREAKING CHANGE: If you have migrated to a plugin-based backup and recovery
solution such as Barman Cloud, these fields are no longer synchronized and will
not be updated. Users still relying on the in-core support for Barman Cloud
and volume snapshots can continue to use these fields for the time being.

Closes #8051

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>