docs(cnpg-i): deprecate backup metrics and fields in `Cluster` status (#8091)
Due to the transition to a plugin-based, backup-agnostic architecture in
CloudNativePG, we are deprecating:

- `firstRecoverabilityPoint`, `firstRecoverabilityPointByMethod`,
`lastSuccessfulBackup`, `lastSuccessfulBackupByMethod`, and
`lastFailedBackup` in `Cluster.status`
- `cnpg_collector_first_recoverability_point`,
`cnpg_collector_last_failed_backup_timestamp`,
`cnpg_collector_last_available_backup_timestamp` metrics

If you have migrated to a plugin-based backup and recovery solution such
as Barman Cloud, these fields/metrics are no longer synchronized and
will not be updated. Users still relying on the in-core support for
Barman Cloud and volume snapshots can continue to use these fields for
the time being.

Closes #8051

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>