feat: introduce experimental support for failover quorum (#7572)
This commit adds opt-in, experimental support for "failover quorum",
also known as "quorum-based failover", in CloudNativePG.

Failover quorum is a mechanism designed to improve data durability and
safety during failover events by ensuring that the promoted replica
contains all synchronously committed data.

With synchronous replication, a transaction is acknowledged only after
all required synchronous standbys have received the WAL data. However,
this alone doesn't guarantee that the operator can always promote the
most advanced replica during a failure.

The failover quorum mechanism addresses this risk by requiring the
operator to verify that a quorum of replicas agrees on the promotion.
This ensures that the selected replica has all required committed data.
If this quorum is not met, failover will not proceed, reducing the risk
of data loss.

To enable the feature, set the annotation:
`cnpg.io/failoverQuorum="true"` in the `Cluster` resource.

Once stabilised, a dedicated configuration field will replace this
annotation.

For further information, refer to the included documentation and the
related issue.

Closes #7481

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>