fix: unblock rollout of former primary Pod when WAL archiving plugin is missing (#8236)
This patch addresses a rollout issue when migrating from the in-tree
Barman Cloud support to the `barman-cloud` plugin, particularly when
using the `switchover` primary update strategy.

During such a migration, a switchover is triggered to roll out the
former primary instance. After the switchover, the former primary Pod
restarts and attempts to archive any remaining WAL files in the queue
before starting PostgreSQL. These files may still be present due to
archiving or misconfiguration delays.

However, the previous version of the Pod does not include support for
the plugin-based archiving system, causing all archiving attempts to
fail during this pre-start phase. As a result, the Pod stalls
indefinitely, blocking the rollout.

This fix introduces a mechanism to detect this specific failure
scenario. If WAL archiving cannot proceed due to a missing plugin, the
Pod exits with a distinct exit code. The operator recognises this exit
code and replaces the Pod with one based on the updated container image,
ensuring the plugin is available and WAL archiving can continue as
expected.

Fixes: #8189

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
(cherry picked from commit 99550adf171efb72d806cea2210fee5698ef5386)