chore(controller_backup): inject VolumeSnapshotReconciler into BackupReconciler
The volumesnapshot.Reconciler was previously created on-the-fly inside the reconcileSnapshotBackup method, worsening the overall performance and making the BackupReconciler harder to test in isolation.

This change refactors the BackupReconciler to receive the volumesnapshot.Reconciler as a dependency on initialization. This improves the design by adhering to dependency injection principles and enhances testability.

The test suite has been updated accordingly to inject the reconciler into the test environment.

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>