fix(backup): do not check Cluster `spec.backup` if method is `plugin` (#8486)
This patch addresses a regression introduced by #8320 that does not
affect any released version.

The regression made it impossible to perform backups using a plugin, as
the controller would incorrectly require and validate the `spec.backup`
field even when the backup method was set to plugin. With this change,
validating the `spec.backup` field is skipped when using plugin-based
backups, restoring the expected behavior for plugin users.

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: theBrahma <office.utpal.brahma@gmail.com>