feat: support loading PostgreSQL extensions via image catalogs
This patch extends imageCatalogs and clusterImageCatalogs CRDs to allow
defining PostgreSQL extensions through the `.spec.images[].extensions` stanza.

This new section uses the same API schema as the Cluster's `.spec.postgresql.extensions`,
providing an additional way to configure image volume extensions using the
image catalog that's being referenced by a Cluster.

Extensions can now be defined at the Cluster's level, inside an image catalog,
or in both. When both sources are present, catalog-defined extensions are merged
with the ones defined at Cluster level, with Cluster configuration taking
precedence for extensions with the same name.
The catalog effectively serves as a default or base layer of configuration,
which the Cluster can override.

Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>