fix(cnpgi,connection): improve retry algorithm (#8554)
On each execution of the plugin reconciliation loop, the plugin’s
connection pool is closed and recreated with the newly detected
settings.

If a connection request occurs while the pool is being recreated, the
request may fail with a "closed pool" error.

This patch improves the reconciliation logic to properly detect pool
changes and enhances the retry mechanism by introducing an exponential
backoff strategy.

Fixes: #8553

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>