chore(cluster): add support for pprof via `alpha.cnpg.io/enableInstancePprof` annotation (#7876)
This change introduces the ability to enable the Go pprof HTTP server in
instance pods by setting the `alpha.cnpg.io/enableInstancePprof`
annotation to "true" on a Cluster resource.

When enabled, each instance pod exposes a pprof server on port 6060/TCP,
making it possible to collect and analyse profiling data for debugging
and performance tuning. The operator automatically manages the rolling
update needed to apply or remove the annotation.

Closes #4056

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>