chore(lint): lint import aliases (#8604)
The main goal of this patch is to eliminate all unqualified `v1`
aliases.

In multiple places within the code, we import API packages that end with
`v1`, all of which are aliased as `v1`. When reading the code,
encountering `v1.XxxYyyZzz(...)` makes it unclear which package's
functions are being invoked.

In contrast, we have a more consistent naming model in other places,
where we use more descriptive names such as `appsv1`, `metav1`, or
`corev1`.

This patch implements a linter for the imports of these packages,
ensuring that they are consistently imported using the longer, more
explanatory names.

A comment in the patch highlights the relevant section.

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>