chore: rename `authQuerySecret` to `serverTLSSecret`
This patch introduces a new field, `.spec.pgbouncer.serverTLSSecret`,
which can reference either a TLS secret or a BasicAuth secret:

- TLS secret: mapped to the `server_tls_key_file` and
  `server_tls_cert_file` parameters in the PgBouncer configuration. These
  files are used by PgBouncer to authenticate to the PostgreSQL server and
  by PostgreSQL to validate the provided certificate.

- BasicAuth secret: provides the username and password that PgBouncer
  will use when running `auth_query` against the PostgreSQL server.

The goal of this change is to align the Pooler CRD naming with the
corresponding PgBouncer settings, simplifying configuration for users
already familiar with PgBouncer.

For backward compatibility, the existing `authQuerySecret` field is
still accepted. The validation webhook now raises a warning to guide
users toward migrating to the new name. If both fields are specified,
`serverTLSSecret` takes precedence.

See: #8675

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>