fix: wait for application user secret before starting initdb (#8663)
This patch ensures that the operator waits for the application user secret
to be available before running `initdb`.

The change is implemented with minimal impact by adding a reference to the
secret in the Pod created by the `initdb` Job. The username is exposed as an
environment variable, which is not currently used, but guarantees that the Job
cannot start until the secret exists.

Closes: #6069

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
(cherry picked from commit 9d4ec08219e050800a5b5eac4625e7f1deab147e)