ci: shorten the matrix.id of E2E jobs (#8725)
Recently we updated the E2E's PostgreSQL version to use tags that
contain the image type and the OS.

As a result, the ID of each job generated by
`.github/e2e-matrix-generator.py` now is even longer and not
compatible with some Cloud Providers that have a maximum length
for Cluster names.  

This patch shortens the ID of the E2e job, resulting in a shorter
cluster name length.

Closes #8714

Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>