feat: "unrecoverable" annotation (#8178)
This patch introduces the annotation `alpha.cnpg.io/unrecoverable=true`,
which can be added to any replica instance Pod.
When this annotation is present, the operator will permanently delete
the associated instance by removing both the PVCs and the Pod itself.
Following this, the operator will recreate the missing instance from the
primary instance.

Please note that this annotation is available only for replica
instances.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>