fix(metrics): add cnpg_cluster additionally to cluster prometheus label
Issue: https://github.com/cloudnative-pg/cloudnative-pg/issues/2501

If this is merged, I will create a follow-up PR for the
grafana-dashboard so that the new `cnpg_cluster` label is used there
instead of the `cluster` label. The old `cluster` label can be deleted
in the future, but we keep it for now.

Signed-off-by: Florian Ströger <stroeger@youniqx.com>