feat(monitoring): cache for monitoring queries (#8003)
To enhance resource efficiency, the outputs of monitoring queries are
now cached. The cache has a configurable expiration time, which
defaults to 30 seconds and can be adjusted using the
`cluster.spec.monitoring.metricsQueriesTTL` field.

To maintain the existing behavior of executing the query on every
endpoint scrape, users should set the `metricsQueriesTTL` field to
zero. Note that when set to zero, the webhook will issue a warning,
indicating that this configuration may lead to overloading the
PostgreSQL server depending on the Prometheus setup.

Closes #4962

Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@mailfence.com>
Co-authored-by: Jeff Mealo <jmealo@protonmail.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>