WIP: Allow postgres major downgrade
Also allows patching the instance with lower version of PG

If a plugin is created which supports downgrade by providing label: downgrade-plugin: "true"
The operator can allow changing the postgres version to a lower one

This gives the plugin a way to enable downgrade of postgres major
version.
The plugin can then remove everything on the file system and initialize
the file struture of the operand.

This is an entirely opt-in feature which does not change the existing
behavior.