chore: sign all the artifacts
We use GoReleaser for the packages, but so far, we were just signing only
a set of artifacts (checksums) and not all of them by default, this lower
the score on the OpenSSF review. Now, by default, we sign all the artifacts
to avoid any missing signature in the future.

Closes #9163

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>