chore: sign all release artifacts (#9166)
We use GoReleaser to create our packages, but until now, we only signed
a limited number of artifacts, specifically the checksums. This led to a
lower score in the OpenSSF review.

With this update, all artifacts will now be signed by default, ensuring
complete coverage and preventing any missing signatures in future
releases.

Closes #9163

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>